backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
httpx==0.24.0
python-dotenv==1.0.0
# Use a redis version available on PyPI for this environment
redis==4.6.0
pytest==7.4.0
pytest-asyncio==0.21.0
python-multipart==0.0.6
# Rate limiting
slowapi==0.1.9
# Pydantic for data validation (v2 compatible with Python 3.13)
pydantic>=2.9.0
